+++ subgraphs/arbitrum/src/contracts/Constants.ts
@@ -39,8 +39,12 @@ export const TREASURE_ATLAS_MINE = "0xa0a89db1c899c49f98e6326b764bafcf167fc2ce".
 
 export const OLYMPUS_LENDER = "0x868C3ae18Fdea85bBb7a303e379c5B7e23b30F03".toLowerCase();
 
+export const SILO_REPOSITORY = "0x8658047e48CC09161f4152c79155Dac1d710Ff0a".toLowerCase();
+
 export const SILO_ADDRESS = "0x9992f660137979C1ca7f8b119Cd16361594E3681".toLowerCase();
-export const SENTIMENT_LTOKEN = "0x3c34c679a76697e12e8a7496173487fea004f7c0".toLowerCase();
+export const SENTIMENT_LTOKEN = "0x37E6a0EcB9e8E5D90104590049a0A197E1363b67".toLowerCase();
+
+export const JONES_WRITE_OFF_BLOCK = "130482707";
 
 export const SILO_DEPLOYMENTS = new Array<LendingMarketDeployment>();
 SILO_DEPLOYMENTS.push(new LendingMarketDeployment(ERC20_OHM, BigInt.fromString("99067079"), BigDecimal.fromString("25000"), SILO_ADDRESS)); // https://arbiscan.io/tx/0x55cabbd6cd41d2fa79a6c93743729bbfa85577ff3e92255f27bfd832344871f6
@@ -179,6 +183,7 @@ CONTRACT_NAME_MAP.set(OLYMPUS_LENDER, "Olympus Lender");
 CONTRACT_NAME_MAP.set(RARI_ALLOCATOR, "Rari Allocator");
 CONTRACT_NAME_MAP.set(SENTIMENT_LTOKEN, "Sentiment Finance");
 CONTRACT_NAME_MAP.set(SILO_ADDRESS, "Silo Router");
+CONTRACT_NAME_MAP.set(SILO_REPOSITORY, "Silo Repository");
 CONTRACT_NAME_MAP.set(TREASURE_ATLAS_MINE, "TreasureDAO Atlas Mine");
 CONTRACT_NAME_MAP.set(TREASURY_ADDRESS_V1, "Treasury Wallet V1");
 CONTRACT_NAME_MAP.set(TREASURY_ADDRESS_V2, "Treasury Wallet V2");