+++ subgraphs/ethereum/src/utils/OhmCalculations.ts
@@ -50,6 +50,7 @@ import {
 import { PairHandlerTypes } from "./PairHandler";
 import { getUSDRate } from "./Price";
 import { getUniswapV3OhmSupply } from "../liquidity/LiquidityUniswapV3";
+import { getSiloSupply } from "./Silo";
 
 const MIGRATION_OFFSET_STARTING_BLOCK = "14381564";
 const MIGRATION_OFFSET = "2013";
@@ -77,6 +78,16 @@ const BLV_INCLUSION_BLOCK = "17620000";
  * the historical values.
  */
 const OLYMPUS_INCUR_DEBT_BLOCK = "17620000";
+
+/**
+ * The block from which the sOHM calculations were corrected to remove the multiplication by index.
+ */
+const SOHM_INDEX_CORRECTION_BLOCK = "18121728";
+
+/**
+ * The block from which the balance of the Silo Borrowable OHM token was used, instead of manual deployments.
+ */
+const SILO_TOKEN_BLOCK = "18121728";
 
 /**
  * Returns the total supply of the latest version of the OHM contract
@@ -371,10 +382,19 @@ function getLendingMarketDeploymentOHMRecords(timestamp: BigInt, deploymentAddre
 export function getMintedBorrowableOHMRecords(timestamp: BigInt, blockNumber: BigInt): TokenSupply[] {
   const records: TokenSupply[] = [];
 
-  pushTokenSupplyArray(
-    records,
-    getLendingMarketDeploymentOHMRecords(timestamp, SILO_ADDRESS, SILO_DEPLOYMENTS, blockNumber),
-  );
+  // Silo
+  if (blockNumber.lt(BigInt.fromString(SILO_TOKEN_BLOCK))) {
+    pushTokenSupplyArray(
+      records,
+      getLendingMarketDeploymentOHMRecords(timestamp, SILO_ADDRESS, SILO_DEPLOYMENTS, blockNumber),
+    );
+  }
+  else {
+    pushTokenSupplyArray(
+      records,
+      getSiloSupply(timestamp, blockNumber),
+    );
+  }
 
   pushTokenSupplyArray(
     records,
@@ -450,7 +470,7 @@ export function getTreasuryOHMRecords(timestamp: BigInt, blockNumber: BigInt): T
       if (balance.equals(BigDecimal.zero())) continue;
 
       // Derive the OHM balance
-      const ohmBalance = ohmIndex.times(balance);
+      const ohmBalance = blockNumber.ge(BigInt.fromString(SOHM_INDEX_CORRECTION_BLOCK)) ? balance : ohmIndex.times(balance);
 
       records.push(
         createOrUpdateTokenSupply(
@@ -603,6 +623,11 @@ export function getIncurDebtSupplyRecords(timestamp: BigInt, blockNumber: BigInt
   // Create a TokenSupply record
   const ohmDecimals = getERC20Decimals(ERC20_OHM_V2, blockNumber);
   const outstandingDebt = toDecimal(outstandingDebtResult.value, ohmDecimals);
+
+  // Ignore zero balance
+  if (outstandingDebt.equals(BigDecimal.zero())) {
+    return records;
+  }
 
   records.push(
     createOrUpdateTokenSupply(