+++ subgraphs/ethereum/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Subgraph Changelog
 
+## 4.12.2 (2023-09-13)
+
+- OHM supply: Fix the incorrect multiplication of sOHM balance by OHM index
+- OHM supply: Going forward, dynamically determine the OHM supply minted into Silo Finance
+- Liquid backing: Write down the value of cvxCRV that is bricked/inaccessible in the Convex allocator contract
+- Liquid backing: Add missing withdrawal from Myso pool
+
 ## 4.11.3 (2023-09-12)
 
 - Adds support for protocol-owned liquidity in Uniswap V3 (both assets and OHM supply)